+++ lib/draftking/cli/commands/blogs.rb
@@ -1,10 +1,11 @@
 module DK
   class CLI < Thor
     desc 'blogs', 'Display a list of blogs under the current account.'
-    option :simulate,      type: :boolean, aliases: :s, desc: Options.op_strings[:simulate]
+    option :simulate, type: :boolean, aliases: :s, desc: Options.op_strings[:simulate]
+    option :config,   type: :string,  desc: Options.op_strings[:config]
     def blogs
       configured?
-      self.class.blogs_print_list(get_dk_instance(options))
+      self.class.blogs_print_list(get_dk_instance(process_options(options)))
     end
 
     private
@@ -12,12 +13,14 @@ module DK
     # Print blog list
     # @param dk [DK::Client] Instance of tumblr_draftking
     def self.blogs_print_list(dk)
-      result = "\n#-------- Blogs --------#"
-      dk.user.blogs.each_with_index do |blog, idx|
-        result += "\n#{idx + 1}. #{blog.name}"
-      end
-      puts result += "\n\n" unless dk.simulate
-      result
+      title  = 'Blogs'
+      fields = %w(# blog_name)
+      rows   = []
+      dk.user.blogs.each_with_index { |blog, idx| rows << [idx + 1, blog.name] }
+      report = Reporter.new(title: title, rows: rows, headings: fields)
+
+      report.show unless dk.simulate
+      report
     end
   end
 end