+++ lib/draftking/cli/commands/comment.rb
@@ -22,6 +22,7 @@ module DK
     option :keep_comments, type: :boolean, aliases: :K, desc: Options.op_strings[:keep_comments]
     option :credit,        type: :boolean, desc: Options.op_strings[:credit], default: true
     option :tags,          type: :boolean, desc: Options.op_strings[:tags],   default: true
+    option :config,        type: :string,  desc: Options.op_strings[:config]
     def comment(comm)
       configured?
       opts = process_options(options)