+++ lib/draftking/drafts.rb
@@ -8,7 +8,7 @@ module DK
     # @return [int] Number of modified posts
     def strip_old_comments(options = {})
       options[:message] = 'Stripping previous comments: '
-      mod_count, mod_posts = post_operation(options) do |post, _|
+      mod_count, _mod_posts = post_operation(options) do |post, _|
         post.changed = true
       end
       mod_count
@@ -16,7 +16,7 @@ module DK
 
     def strip_tags(options = {})
       options[:message] = 'Stripping previous comments: '
-      mod_count, mod_posts = post_operation(options) do |post, _|
+      mod_count, _mod_posts = post_operation(options) do |post, _|
         post.clear_tags
       end
       mod_count
@@ -37,7 +37,7 @@ module DK
       options[:shuffle] = true
       options[:state]   = DK::QUEUE
       options[:limit] ||= @q_space
-      mod_count, mod_posts = post_operation(options) do |post, index|
+      mod_count, _mod_posts = post_operation(options) do |post, index|
         next false unless index_within_limit?(index, @q_space)
         next false unless post.has_key_text?(@key_text)
         post.replace_comment_with(@comment)